add option to include config in lambda package
the config in SSM in us-east-1 can be slow to access, which causes
sporadic 503 errors in regions far from us-east-1. This change adds a
variable lambda_ship_config which defaults to false.  If this variable
is set to true, the json config will instead be written to a file and
packaged/deployed with the lambda code - see the comments in the README
for further rationale.

default behaviour is to continue to use SSM, so this should be a
non-breaking change.